refactor: moved authorLabel selection logic to utils
